Removed setting of SSL flags via bindings.
SSL flags do not belong on the binding, they should be applied at the site
level. Also implementation did not account for the varying return types of
the Get-WebBinding cmdlet. This resulted in the Test-TargetResource always
returning false for sites that did not have SSL flags set.

The ability to set SSL flags is not required at the moment and will not be
re-introduced in the forseeable future.