Fix for issue #4
Array index exception when updating web site bindings
https://github.com/SEEK-Jobs/DSC/issues/4

Interpretation of BindingInformation for net.tcp and net.pipe bindings was
broken. Comparison function would split the BindingInformation using ':'
as a delimiter and assumed that there would always be three elements
returned. This is not the case for net.tcp and net.pipe bindings.

Replaced comparison logic with protocol-specific regular expressions and
added tests to verify behaviour.