Fixes issue #8
Each binding was being individulally applied to IIS in a tight loop. This
was increasing the likelihood of encountering a lock on the
applicationHost.config.xml file. The implementation has been changed to
apply all bindings at once. It is still however necessary to apply
certificates to https bindings separately.