cAppPool ApplicationName is now obsolete
The ApplicationName was added as a key to the AppPool as a workaround in
the scenario where two website shared the same application pool. This was
a bad idea. Two DSC resources that manipulate the state of the same IIS
resource can potentially yield unpredicable results.

This change will impact DSC configuration creating multiple application
pools with the same name, resulting in the following error:

"The key properties combination 'AppPool' is duplicated for keys 'Name'
of resource 'cWebAppPool'"

The fix is to create a single instance of the application pool before the
web sites that depend on it.