Bindings resource derives SSL cert thumbprint.
It is now unnecessary to provide certificate thumbprint and store names as part
of an SSL binding.

When binding a certificate to an endpoint (ip:port), it will clobber any
existing binding to that endpoint.

Removes https binding comparison which was not working anyway. A working
comparison may need to be added in the future.