Assume empty IP address bindings are for all IPs
NOTE: we still no not fully support detecting conflicts between sites
that use wildcard IP addresses. If two sites try and assign a port for
all IP addresses, this will be detected as a conflict. However if one
site already has assigned all IP addresses for a given port and a
binding is added for a new site that binds to the same port but on a
specific IP address; this will not be detected as a conflict.